// src/app/shared/model/Biker.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Biker {
  aniExperienta?: number;

  bicicleta?: string;

  contact?: models.Contact;

  dataNasterii?: Date;

  disciplinaFavorita?: string;

  id?: number;

  membruData?: Date;

  nume?: string;

  prenume?: string;

  user?: models.User;

}